chore(dev): stop route files from acting as refresh boundaries and unloop the routes watcher
Frontend: MobileRoutes.tsx, Routes.tsx, DrillDownModal.tsx and
EmbedCodeSnippet.tsx exported non-components next to components, so every
HMR update reaching them (every edit to @lightdash/common) hit "Could not
Fast Refresh" and re-rendered the app from the root. Move MobileNavBar,
RedirectToResource, NavBarLayout and combineFilters into their own files
and stop exporting the SnippetLanguage enum.

pm2: the routes watcher ran the root generate-api-dev script, whose extra
common-build duplicates common-watch and crash-looped at 1Hz on a stale
incremental build. Call the backend script directly and back off on
failure. Also ignore backend test files in the API watch list. Apply the
same three changes to the agent-harness template so both harnesses keep
one contract, and update the PM2 harness test that pins it.
